+++ CHANGELOG.md
@@ -150,6 +150,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - uniform shebang usage in scripts of init.d folder #821
 - Check if commas are allowes in node/profile set
 - Removed paths.tftpdir and paths.datadir
+- Updated WW_INTERNAL to 45 for upcoming 4.5.x release
 
 ## [4.4.0] 2023-01-18